Move SocialMediaLinksStyles to relevant component
